+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 - Major: Added 2 new slots (`Homies badges`) & Changed the host of the main slot to `https://chatterinohomies.com`
 - Major: Added /nuke command
+- Minor: `F5` now reloads the badges
 
 ## Homies (#3)
 - Major: Added Homies emotes (Global & Channel emotes)